fix: lesson price to int

+++ SPA/Application/Lessons/Commands/CreateLessonCommand/CreateLessonCommand.cs
@@ -5,4 +5,4 @@ using SPA.Domain;
 
 namespace SPA.Application.Lessons.Commands.CreateLessonCommand;
 
-internal sealed record CreateLessonCommand(Guid TutorId, DateTimeOffset Start, DateTimeOffset End, double Price, LessonType Type) : IRequest<Lesson?>;
+internal sealed record CreateLessonCommand(Guid TutorId, DateTimeOffset Start, DateTimeOffset End, int Price, LessonType Type) : IRequest<Lesson?>;
