+++ SPA/Domain/Lesson.cs
@@ -4,7 +4,7 @@ public sealed class Lesson
 {
     public Guid Id { get; init; }
 
-    public double Price { get; init; }
+    public int Price { get; init; }
 
     public LessonStatus Status { get; init; }
